\name{lucaDat}
\alias{lucaDat}
\docType{data}
\title{Simulated data for a hypothetical binary trait}
\description{
Simulated data used to illustrate the luca package.
} 
\usage{data(lucaDat)}
\format{A data frame with 1000 observations on the following 4 variables:
\tabular{rlll}{
[,1] \tab d \tab numeric \tab disease status (1=yes, 0=no) \cr
[,2] \tab g \tab factor \tab genetic factor with levels \code{A/A}, \code{A/C}, \code{C/C} \cr
[,3] \tab a \tab numeric \tab first non-genetic attribute \cr
[,4] \tab a2 \tab numeric \tab second non-genetic attribute }
}
\examples{
data(lucaDat)
}
\keyword{datasets}
